A1/Solutions/requirements.txt
nltk
pyspellchecker
wordcloud
matplotlib